requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
backports.tarfile==1.1.0
bitarray==2.9.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
cytoolz==0.12.3
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
hexbytes==1.2.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jaraco.context==5.3.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
more-itertools==10.2.0
multidict==6.1.0
numpy==1.24.4
packaging==23.2
pandas==2.0.3
parsimonious==0.10.0
pillow==10.4.0
py-solc-x==2.0.3
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyunormalize==15.1.0
regex==2024.7.24
requests==2.31.0
rlp==4.0.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
SpeechRecognition==3.10.3
threadpoolctl==3.5.0
toolz==0.12.1
types-requests==2.32.0.20240907
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
web3==7.2.0
websockets==13.0.1
Werkzeug==3.0.3
wolframalpha==5.0.0
xmltodict==0.13.0
yarl==1.11.1
zipp==3.19.2
